move js code to utils component
